pyproject.toml

[project]
name = "tickertrending-bot"
version = "0.1.0"
description = "Telegram bot to track and boost Ethereum, Solana, and BNB tokens"
authors = ["Aras Baker <arasbaker8@gmail.com>"]
requires-python = ">=3.11"

dependencies = [
    "aiohttp>=3.8.6",
    "python-telegram-bot>=22.0",
    "requests>=2.31.0",
    "web3>=6.10.0",
    "python-dotenv>=1.0.1"
]

[tool.setuptools]
packages = ["."]
